analytics/terraform/spark-k8s-operator/mwaa222/requirements.txt
kubernetes==11.0.0
apache-airflow-providers-cncf-kubernetes==2.0.2
